Always use \xHH instead of just \HH in regexes.
Recent perl (5.22.x) complains about uses of \XX.  In fact it thinks
that \2D is a $2 followed by a literal D, and emits warnings.  (I did
not actually build test cases to confirm if it was evaluating it
incorrectly at run time).  \00 should be unambiguous, but I went
ahead and switched to \xHH in all cases, for consistency.